requirements.txt
flask==3.0.2
nltk==3.8.1
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0
SpeechRecognition==3.10.1
numpy==1.26.4
pandas==2.2.1
spacy==3.7.4
python-dotenv==1.0.1
Werkzeug==3.0.1
joblib==1.3.2 
pydub==0.25.1
soundfile==0.12.1
transformers==4.38.2
torch==2.2.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl 
gunicorn 